removed types.ts from search cause it was unnecessary
