+++ src/search/searchContainer.ts
@@ -1,4 +1,5 @@
 import { connect } from 'react-redux'
+import { ActionCreator } from 'redux';
 import { ThunkAction, ThunkDispatch } from 'redux-thunk'
 import { GET_COMPANY_ACTION } from '../companyReducer'
 import { getHistoryData } from '../history'
@@ -7,18 +8,20 @@ import { getNewsData } from '../news'
 import { getOverviewData } from '../overview'
 import { getPeersData } from '../peers'
 import { API, GlobalState } from '../types'
+import { Company } from '../types'
 import Search from './search'
-import { Company, SearchContainerAction } from './types'
 
-type ThunkResult<R> = ThunkAction<R, GlobalState, API, SearchContainerAction>
+export interface SearchContainerAction { type: string, company: Company }
+
 
 const mapStateToProps = (state: GlobalState) => ({
   companySymbols: state.companySymbols
 })
 
-const getCompanyAction = (company: Company): SearchContainerAction => ({ type: GET_COMPANY_ACTION, company })
+const getCompanyAction: ActionCreator<SearchContainerAction> = (company: Company) => ({ type: GET_COMPANY_ACTION, company })
 
-export const getCompanyInfo: (company: Company) => ThunkResult<void> = company => (dispatch) => {
+type ThunkResult = ThunkAction<void, GlobalState, API, SearchContainerAction>
+export const getCompanyInfo: (company: Company) => ThunkResult = company => (dispatch) => {
   dispatch(getHistoryData(company.symbol))
   dispatch(getNewsData(company.symbol))
   dispatch(getKeyStatsData(company.symbol))
@@ -27,7 +30,8 @@ export const getCompanyInfo: (company: Company) => ThunkResult<void> = company =
   dispatch(getCompanyAction(company))
 }
 
-const mapDispatchToProps = (dispatch: ThunkDispatch<GlobalState, API, SearchContainerAction>) => ({
+type ThunkDispatchContainerAction = ThunkDispatch<GlobalState, API, SearchContainerAction>
+const mapDispatchToProps = (dispatch: ThunkDispatchContainerAction) => ({
   getInfo: (company: Company) => dispatch(getCompanyInfo(company))
 })
