--- src/search/types.ts
@@ -1,12 +1,0 @@
-export interface Company {
-  name: string
-  symbol: string
-}
-export type CompanySelect = (company: Company) => void
-
-export interface SearchProps {
-  readonly getInfo: CompanySelect
-  readonly companySymbols: Company[]
-}
-
-export interface SearchContainerAction { type: string, company: Company }
